STY: moved test utility
Moved a useful test utility to a separate file to allow use by multiple test classes.
