+++ aetherpy/tests/utils.py
@@ -1,0 +1,29 @@
+#!/usr/bin/env python
+# Copyright 2021, the Aether Development Team (see doc/dev_team.md for members)
+# Full license can be found in License.md
+"""Utilities for test classes."""
+
+import os
+
+
+def sys_agnostic_rename(inname, outname, max_attemps=100):
+    """Wrap os.rename, Windows OS sometimes needs time to allow rename to work.
+
+    Parameters
+    ----------
+    inname : str
+        Input filename or directory
+    outname : str
+        Output filename or directory
+    max_attemps : int
+        Maximum rename attemps (default=100)
+
+    """
+    for retry in range(max_attemps):
+        try:
+            os.rename(inname, outname)
+            break
+        except Exception:
+            pass
+
+    return
